refs #121 - refactors notification handling and add test code
